// App/src/app/models/player-interface.ts
export interface PlayerInterface {
    PlayerID: number;
    Position: string;
    FirstName: string;
    LastName: string;
    Country: string;
    Appearances: number;
    Goals: number;
    Assists: number;
    CleanSheets: number;
    OnLoan: string;
    LoanTeam: string;
    ImageName: string;
    TotalPlayers: number;
}